+++ Skillet-Classic/SkilletNewsData.lua
@@ -26,6 +26,18 @@ local isCata = WOW_PROJECT_ID == WOW_PROJECT_CATACLYSM_CLASSIC
 
 Skillet.NewsName = "Skillet-Classic News"
 Skillet.NewsData = {
+	{	version = "2.18",
+		data = {
+			{	name = "Changes",
+				data = {
+					{ header = "Searching", body = "Update Tooltip data collection\nUse '+' in Search for 'and'" },
+					{ header = "Enchanting", body = "Update Detail Buttons on a per recipe basis" },
+					{ header = "Packaging", body = "Remove LibDialog" },
+					{ header = "TOC", body = "Update TOC" },
+				},
+			},
+		},
+	},
 	{	version = "2.17",
 		data = {
 			{	name = "Changes",